e2e: extract shared route-convergence diagnostics helper across IBRL tests
