+++ e2e/multi_client_ibrl_liveness_test.go
@@ -327,27 +327,9 @@ func runMultiClientIBRLRouteLivenessTest(t *testing.T, log *slog.Logger, dn *dev
 
 	// Wait for clients to have routes to each other before starting liveness tests.
 	log.Debug("==> Waiting for client routes before liveness testing")
-	require.Eventually(t, func() bool {
-		output, err := client1.Exec(t.Context(), []string{"ip", "r", "list", "dev", "doublezero0"})
-		if err != nil {
-			return false
-		}
-		return strings.Contains(string(output), client2DZIP) && strings.Contains(string(output), client3DZIP)
-	}, 60*time.Second, 1*time.Second, "client1 should have routes to client2 and client3")
-	require.Eventually(t, func() bool {
-		output, err := client2.Exec(t.Context(), []string{"ip", "r", "list", "dev", "doublezero0"})
-		if err != nil {
-			return false
-		}
-		return strings.Contains(string(output), client1DZIP)
-	}, 60*time.Second, 5*time.Second, "client2 should have route to client1")
-	require.Eventually(t, func() bool {
-		output, err := client3.Exec(t.Context(), []string{"ip", "r", "list", "dev", "doublezero0"})
-		if err != nil {
-			return false
-		}
-		return strings.Contains(string(output), client1DZIP)
-	}, 60*time.Second, 5*time.Second, "client3 should have route to client1")
+	requireClientHasRoutes(t, log, "client1", client1, 60*time.Second, 1*time.Second, client2DZIP, client3DZIP)
+	requireClientHasRoutes(t, log, "client2", client2, 60*time.Second, 5*time.Second, client1DZIP)
+	requireClientHasRoutes(t, log, "client3", client3, 60*time.Second, 5*time.Second, client1DZIP)
 	log.Debug("--> Client routes established")
 
 	// --- Route liveness block matrix ---
